worker/requirements.txt
httpx==0.27.0
motor==3.6.0
python-dotenv==1.0.1
transformers==4.46.0
torch==2.5.0
